Split router into router anf factory
