+++ softgroup/basis/common-router/src/main/java/com/softgroup/common/factory/HandlerFactory.java
@@ -1,0 +1,11 @@
+package com.softgroup.common.factory;
+
+import com.softgroup.common.protocol.Request;
+import com.softgroup.common.router.api.Handler;
+
+/**
+ * Created by maxim on 06.03.17.
+ */
+public interface HandlerFactory {
+    Handler getHandler(Request<?> msg);
+}
