+++ softgroup/basis/common-router/src/main/java/com/softgroup/common/factory/TypeHandlerFactory.java
@@ -1,0 +1,14 @@
+package com.softgroup.common.factory;
+
+import com.softgroup.common.protocol.Request;
+import com.softgroup.common.router.api.Handler;
+
+/**
+ * Created by user on 06.03.2017.
+ */
+public class TypeHandlerFactory<T extends Handler> extends AbstractHandlerFactory<T> implements HandlerFactory{
+
+    public String getRouteKey(Request<?> msg) {
+        return msg.getHeader().getType();
+    }
+}
