+++ softgroup/basis/common-router/src/main/java/com/softgroup/common/router/api/AbstractRouterHandler.java
@@ -1,31 +1,19 @@
 package com.softgroup.common.router.api;
 
+import com.softgroup.common.factory.HandlerFactory;
 import com.softgroup.common.protocol.Request;
 import com.softgroup.common.protocol.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import javax.annotation.PostConstruct;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-public abstract class AbstractRouterHandler<T extends Handler> implements RouterHandler {
+public abstract class AbstractRouterHandler<T extends HandlerFactory> implements RouterHandler {
 
 	@Autowired
-	private List<T> handlers;
-	private Map<String,T> handlerMap = new HashMap<>();
+	private T factory;
 
 	@Override
 	public Response<?> handle(Request<?> msg) {
-		T handler = handlerMap.get(getRouteKey(msg));
-		return handler.handle(msg);
+		return factory.getHandler(msg).handle(msg);
 	}
 
 
-	@PostConstruct
-	public void init () {
-		for (T handler:handlers) {
-			handlerMap.put(handler.getName(),handler);
-		}
-	}
 }
