+++ softgroup/basis/common-router/src/main/java/com/softgroup/common/router/api/RouterHandler.java
@@ -1,8 +1,6 @@
 package com.softgroup.common.router.api;
 
-import com.softgroup.common.protocol.Request;
 
 public interface RouterHandler extends Handler {
 
-	String getRouteKey(final Request<?> msg);
 }
