+++ softgroup/services/authorization/authorization-api/src/main/java/com/softgroup/authorization/api/factory/AuthorizationHandlerFactory.java
@@ -1,0 +1,9 @@
+package com.softgroup.authorization.api.factory;
+
+import com.softgroup.common.factory.HandlerFactory;
+
+/**
+ * Created by user on 06.03.2017.
+ */
+public interface AuthorizationHandlerFactory extends HandlerFactory {
+}
