+++ softgroup/services/authorization/authorization-impl/src/main/java/com/softgroup/authorization/impl/router/AuthorizationRouter.java
@@ -1,7 +1,7 @@
 package com.softgroup.authorization.impl.router;
 
-import com.softgroup.authorization.api.router.AuthorizationRequestHandler;
-import com.softgroup.common.router.api.AbstractCommandRouterHandler;
+import com.softgroup.authorization.api.factory.AuthorizationHandlerFactory;
+import com.softgroup.common.router.api.AbstractRouterHandler;
 import com.softgroup.common.router.api.CommandRouterHandler;
 import org.springframework.stereotype.Component;
 
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
  * Created by user on 24.02.2017.
  */
 @Component
-public class AuthorizationRouter extends AbstractCommandRouterHandler<AuthorizationRequestHandler> implements CommandRouterHandler {
+public class AuthorizationRouter extends AbstractRouterHandler<AuthorizationHandlerFactory> implements CommandRouterHandler {
 
     @Override
     public String getName(){
