+++ softgroup/services/messenger/messenger-api/src/main/java/com/softgroup/messenger/api/factory/MessengerHandlerFactory.java
@@ -1,0 +1,9 @@
+package com.softgroup.messenger.api.factory;
+
+import com.softgroup.common.factory.HandlerFactory;
+
+/**
+ * Created by user on 06.03.2017.
+ */
+public interface MessengerHandlerFactory extends HandlerFactory {
+}
