+++ softgroup/services/messenger/messenger-impl/src/test/java/com/softgroup/messenger/impl/router/MessengerRouterUnitTest.java
@@ -1,6 +1,8 @@
 package com.softgroup.messenger.impl.router;
 
+import com.softgroup.common.factory.AbstractHandlerFactory;
 import com.softgroup.common.protocol.*;
+import com.softgroup.common.router.api.Handler;
 import com.softgroup.messenger.api.router.MessengerRequestHandler;
 import org.hamcrest.CoreMatchers;
 import org.junit.Before;
@@ -17,9 +19,8 @@ import java.util.HashMap;
 import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Matchers.anyObject;
+import static org.mockito.Mockito.*;
 
 /**
  * Created by user on 03.03.2017.
@@ -30,72 +31,24 @@ import static org.mockito.Mockito.when;
         @InjectMocks
         private MessengerRouter router;
 
-        @Spy
-        private ArrayList<MessengerRequestHandler> handlers;
-        @Spy
-        private HashMap<String,MessengerRequestHandler> handlersMap;
         @Mock
-        private MessengerRequestHandler createConversationHandler;
+        private AbstractHandlerFactory factory;
+
         @Mock
-        private MessengerRequestHandler deleteConversationHandler;
+        private Handler handler;
 
-        @Test
-        public void isRouterExist(){
-            assertThat(router, CoreMatchers.notNullValue());
-        }
+        @Mock
+        Request request;
 
-        private Request<RequestData> createConversationRequestREST;
-        private Request<RequestData> deleteConversationRequestREST;
-        private Response createConversationResponseREST;
-        private Response deleteConversationResponseREST;
         @Before
-        public void createRequestResponse(){
-            createConversationRequestREST = new Request<RequestData>();
-            ActionHeader header = new ActionHeader();
-            header.setCommand("create_conversation");
-            header.setType("messenger");
-            createConversationRequestREST.setHeader(header);
-
-
-            createConversationResponseREST = new Response<ResponseData>();
-
-            when(createConversationHandler.handle(createConversationRequestREST)).thenReturn(createConversationResponseREST);
-            when(createConversationHandler.getName()).thenReturn("create_conversation");
-
-            deleteConversationRequestREST = new Request<RequestData>();
-            ActionHeader header1 = new ActionHeader();
-            header1.setCommand("delete_conversation");
-            header1.setType("messenger");
-            deleteConversationRequestREST.setHeader(header1);
-
-            deleteConversationResponseREST = new Response<ResponseData>();
-
-            when(deleteConversationHandler.handle(deleteConversationRequestREST)).thenReturn(deleteConversationResponseREST);
-            when(deleteConversationHandler.getName()).thenReturn("delete_conversation");
-
-            handlers.add(createConversationHandler);
-            handlers.add(deleteConversationHandler);
-
-            for(MessengerRequestHandler h:handlers){
-                handlersMap.put(h.getName(),h);
-            }
-
+        public void prepareFactory(){
+            when(factory.getHandler(anyObject())).thenReturn(handler);
         }
 
         @Test
-        public void traceRouteToCreateConversation(){
-            Response r = router.handle(createConversationRequestREST);
-            assertThat(r,is(instanceOf(Response.class)));
-            verify(createConversationHandler).handle(createConversationRequestREST);
-            verify(deleteConversationHandler,never()).handle(createConversationRequestREST);
-        }
-
-        @Test
-        public void traceRouteToDeleteConversation(){
-            Response r = router.handle(deleteConversationRequestREST);
-            assertThat(r,is(instanceOf(Response.class)));
-            verify(deleteConversationHandler).handle(deleteConversationRequestREST);
-            verify(createConversationHandler,never()).handle(deleteConversationRequestREST);
+        public void routerHandleHandler(){
+            router.handle(request);
+            verify(handler,times(1)).handle(request);
         }
