+++ softgroup/services/profile/profile-api/src/main/java/com/softgroup/profile/api/factory/ProfileHandlerFactory.java
@@ -1,0 +1,9 @@
+package com.softgroup.profile.api.factory;
+
+import com.softgroup.common.factory.HandlerFactory;
+
+/**
+ * Created by user on 06.03.2017.
+ */
+public interface ProfileHandlerFactory extends HandlerFactory {
+}
