+++ softgroup/services/profile/profile-impl/src/main/java/com/softgroup/profile/impl/router/ProfileRouter.java
@@ -1,8 +1,8 @@
 package com.softgroup.profile.impl.router;
-import com.softgroup.common.router.api.AbstractCommandRouterHandler;
 
+import com.softgroup.common.router.api.AbstractRouterHandler;
 import com.softgroup.common.router.api.CommandRouterHandler;
-import com.softgroup.profile.api.router.ProfileRequestHandler;
+import com.softgroup.profile.api.factory.ProfileHandlerFactory;
 import org.springframework.stereotype.Component;
 
 
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
  * Created by user on 26.02.2017.
  */
 @Component
-public class ProfileRouter extends AbstractCommandRouterHandler<ProfileRequestHandler> implements CommandRouterHandler {
+public class ProfileRouter extends AbstractRouterHandler<ProfileHandlerFactory> implements CommandRouterHandler {
 
     @Override
     public String getName(){
