+++ softgroup/services/type-router/type-router-api/src/main/java/com/sofrgroup/router/type/api/factory/TypeRouterFactory.java
@@ -1,0 +1,9 @@
+package com.sofrgroup.router.type.api.factory;
+
+import com.softgroup.common.factory.HandlerFactory;
+
+/**
+ * Created by user on 06.03.2017.
+ */
+public interface TypeRouterFactory extends HandlerFactory {
+}
