+++ softgroup/services/type-router/type-router-impl/src/main/java/com/softgroup/router/type/impl/router/TypeRouter.java
@@ -1,15 +1,15 @@
 package com.softgroup.router.type.impl.router;
 
 import com.sofrgroup.router.type.api.TypeRouterHandler;
-import com.softgroup.common.router.api.AbstractTypeRouterHandler;
-import com.softgroup.common.router.api.CommandRouterHandler;
+import com.sofrgroup.router.type.api.factory.TypeRouterFactory;
+import com.softgroup.common.router.api.AbstractRouterHandler;
 import org.springframework.stereotype.Component;
 
 /**
  * Created by user on 26.02.2017.
  */
 @Component
-public class TypeRouter extends AbstractTypeRouterHandler<CommandRouterHandler>  implements TypeRouterHandler{
+public class TypeRouter extends AbstractRouterHandler<TypeRouterFactory> implements TypeRouterHandler{
     @Override
     public String getName(){
         return "common";
